src/main/java/com/luxoft/movieLand/dao/jdbc/mapper/MovieRowMapper.java
package com.luxoft.movieland.dao.jdbc.mapper;

import com.luxoft.movieland.entity.Movie;
import org.springframework.jdbc.core.RowMapper;
import java.sql.ResultSet;
import java.sql.SQLException;

public class MovieRowMapper implements RowMapper<Movie> {
    @Override
    public Movie mapRow(ResultSet resultSet, int i) throws SQLException {
        Movie movie = new Movie();
        movie.setId(resultSet.getInt("id"));
        movie.setName(resultSet.getString("name"));
        movie.setEnglishName(resultSet.getString("english_name"));
        movie.setYear(resultSet.getInt("year"));
        movie.setDescription(resultSet.getString("description"));
        movie.setRating(resultSet.getDouble("rating"));
        return movie;
    }
}
